family_project/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-extensions==3.2.1
django-rest-framework==0.1.0
django-shell-plus==1.1.7
djangorestframework==3.14.0
psycopg2-binary==2.9.4
python-dotenv==0.21.0
pytz==2022.4
sqlparse==0.4.3
gunicorn==20.1.0
